requirements.txt
requests>=2.28
numpy>=1.24
scipy>=1.10
pandas>=2.0
pyyaml>=6.0
